// packages/nouns-world-engine/src/engine/types.ts
export enum Tile {
  Water = 0,
  Sand = 1,
  Grass = 2,
  Tree = 3,
  Flower = 4,
  Path = 5,
  Rock = 6,
  Spawn = 7,
  DeepWater = 8,
  Arena = 9,
}
// ...
export const TILE_SIZE = 16;
export const MAP_SIZE = 64;